requirements.txt
flask
requests
mediapipe
opencv-python
numpy
google-generativeai
python-dotenv
